refactor and move certain p2p modules and types
