+++ mm2src/coins/Cargo.toml
@@ -67,8 +67,9 @@ mm2_event_stream = { path = "../mm2_event_stream" }
 mm2_git = { path = "../mm2_git" }
 mm2_io = { path = "../mm2_io" }
 mm2_metrics = { path = "../mm2_metrics" }
-mm2_net = { path = "../mm2_net", features = ["p2p"] }
+mm2_net = { path = "../mm2_net" }
 mm2_number = { path = "../mm2_number"}
+mm2_p2p = { path = "../mm2_p2p" }
 mm2_rpc = { path = "../mm2_rpc" }
 mm2_state_machine = { path = "../mm2_state_machine" }
 mocktopus = "0.8.0"
