+++ mm2src/mm2_main/src/lp_native_dex.rs
@@ -30,11 +30,11 @@ use mm2_err_handle::common_errors::InternalError;
 use mm2_err_handle::prelude::*;
 use mm2_event_stream::behaviour::{EventBehaviour, EventInitStatus};
 use mm2_libp2p::behaviours::atomicdex::{generate_ed25519_keypair, GossipsubConfig, DEPRECATED_NETID_LIST};
+use mm2_libp2p::network_event::NetworkEvent;
+use mm2_libp2p::p2p_ctx::P2PContext;
 use mm2_libp2p::{spawn_gossipsub, AdexBehaviourError, NodeType, RelayAddress, RelayAddressError, SeedNodeInfo,
                  SwarmRuntime, WssCerts};
 use mm2_metrics::mm_gauge;
-use mm2_net::network_event::NetworkEvent;
-use mm2_net::p2p::P2PContext;
 use rpc_task::RpcTaskError;
 use serde_json as json;
 use std::convert::TryInto;
