+++ mm2src/mm2_main/src/lp_ordermatch/orderbook_depth.rs
@@ -1,9 +1,10 @@
-use super::{orderbook_topic_from_base_rel, OrdermatchContext, OrdermatchRequest};
-use crate::lp_network::{request_any_relay, P2PRequest};
+use super::{orderbook_topic_from_base_rel, OrdermatchContext};
+use crate::lp_network::request_any_relay;
 use coins::is_wallet_only_ticker;
 use common::log;
 use http::Response;
 use mm2_core::mm_ctx::MmArc;
+use mm2_net::p2p::request_response::{ordermatch::OrdermatchRequest, P2PRequest};
 use serde_json::{self as json, Value as Json};
 use std::collections::HashMap;
