+++ mm2src/mm2_main/src/ordermatch_tests.rs
@@ -7,9 +7,9 @@ use db_common::sqlite::rusqlite::Connection;
 use futures::{channel::mpsc, StreamExt};
 use mm2_core::mm_ctx::{MmArc, MmCtx};
 use mm2_libp2p::behaviours::atomicdex::generate_ed25519_keypair;
+use mm2_libp2p::p2p_ctx::P2PContext;
 use mm2_libp2p::AdexBehaviourCmd;
 use mm2_libp2p::{decode_message, PeerId};
-use mm2_net::p2p::P2PContext;
 use mm2_test_helpers::for_tests::mm_ctx_with_iguana;
 use mocktopus::mocking::*;
 use rand::{seq::SliceRandom, thread_rng, Rng};
