+++ mm2src/mm2_main/src/rpc/lp_commands/lp_commands_legacy.rs
@@ -25,8 +25,8 @@ use common::{rpc_err_response, rpc_response, HyRes};
 use futures::compat::Future01CompatExt;
 use http::Response;
 use mm2_core::mm_ctx::MmArc;
+use mm2_libp2p::p2p_ctx::P2PContext;
 use mm2_metrics::MetricsOps;
-use mm2_net::p2p::P2PContext;
 use mm2_number::construct_detailed;
 use mm2_rpc::data::legacy::{BalanceResponse, CoinInitResponse, Mm2RpcResult, MmVersionResponse, Status};
 use serde_json::{self as json, Value as Json};
