+++ mm2src/mm2_net/Cargo.toml
@@ -7,11 +7,9 @@ edition = "2018"
 doctest = false
 
 [features]
-event-stream = ["mm2_event_stream", "async-stream" , "p2p"]
-p2p = ["mm2-libp2p", "parking_lot"]
 
 [dependencies]
-async-stream = { version = "0.3", optional = true }
+async-stream = { version = "0.3" }
 async-trait = "0.1"
 bytes = "1.1"
 cfg-if = "1.0"
@@ -24,9 +22,7 @@ http = "0.2"
 lazy_static = "1.4"
 mm2_core = { path = "../mm2_core" }
 mm2_err_handle = { path = "../mm2_err_handle" }
-mm2_event_stream = { path = "../mm2_event_stream", optional = true }
-mm2-libp2p = { path = "../mm2_p2p", package = "mm2_p2p", optional = true }
-parking_lot = { version = "0.12.0", features = ["nightly"], optional = true }
+mm2_number = { path = "../mm2_number" }
 prost = "0.12"
 rand = { version = "0.7", features = ["std", "small_rng", "wasm-bindgen"] }
 serde = "1"
