+++ mm2src/mm2_net/src/p2p/request_response/network_info.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Deserialize, Eq, PartialEq, Serialize)]
+pub enum NetworkInfoRequest {
+    /// Get MM2 version of nodes added to stats collection
+    GetMm2Version,
+}
