+++ mm2src/mm2_p2p/src/network_event.rs
@@ -1,12 +1,11 @@
-use crate::p2p::P2PContext;
 use async_trait::async_trait;
 use common::{executor::{SpawnFuture, Timer},
              log::info};
 use futures::channel::oneshot::{self, Receiver, Sender};
+
 use mm2_core::mm_ctx::MmArc;
 pub use mm2_event_stream::behaviour::EventBehaviour;
 use mm2_event_stream::{behaviour::EventInitStatus, Event, EventName, EventStreamConfiguration};
-use mm2_libp2p::behaviours::atomicdex;
 use serde_json::json;
 
 pub struct NetworkEvent {
@@ -22,7 +21,7 @@ impl EventBehaviour for NetworkEvent {
     fn event_name() -> EventName { EventName::NETWORK }
 
     async fn handle(self, interval: f64, tx: oneshot::Sender<EventInitStatus>) {
-        let p2p_ctx = P2PContext::fetch_from_mm_arc(&self.ctx);
+        let p2p_ctx = crate::p2p_ctx::P2PContext::fetch_from_mm_arc(&self.ctx);
         let mut previously_sent = json!({});
 
         tx.send(EventInitStatus::Success).unwrap();
@@ -30,11 +29,11 @@ impl EventBehaviour for NetworkEvent {
         loop {
             let p2p_cmd_tx = p2p_ctx.cmd_tx.lock().clone();
 
-            let directly_connected_peers = atomicdex::get_directly_connected_peers(p2p_cmd_tx.clone()).await;
-            let gossip_mesh = atomicdex::get_gossip_mesh(p2p_cmd_tx.clone()).await;
-            let gossip_peer_topics = atomicdex::get_gossip_peer_topics(p2p_cmd_tx.clone()).await;
-            let gossip_topic_peers = atomicdex::get_gossip_topic_peers(p2p_cmd_tx.clone()).await;
-            let relay_mesh = atomicdex::get_relay_mesh(p2p_cmd_tx).await;
+            let directly_connected_peers = crate::get_directly_connected_peers(p2p_cmd_tx.clone()).await;
+            let gossip_mesh = crate::get_gossip_mesh(p2p_cmd_tx.clone()).await;
+            let gossip_peer_topics = crate::get_gossip_peer_topics(p2p_cmd_tx.clone()).await;
+            let gossip_topic_peers = crate::get_gossip_topic_peers(p2p_cmd_tx.clone()).await;
+            let relay_mesh = crate::get_relay_mesh(p2p_cmd_tx).await;
 
             let event_data = json!({
                 "directly_connected_peers": directly_connected_peers,
