+++ mm2src/mm2_p2p/src/p2p_ctx.rs
@@ -1,10 +1,9 @@
+use libp2p::{identity::Keypair, PeerId};
 use mm2_core::mm_ctx::MmArc;
-use mm2_libp2p::behaviours::atomicdex::AdexCmdTx;
-use mm2_libp2p::PeerId;
 use parking_lot::Mutex;
 use std::sync::Arc;
 
-pub use mm2_libp2p::Keypair;
+use crate::AdexCmdTx;
 
 pub struct P2PContext {
     /// Using Mutex helps to prevent cloning which can actually result to channel being unbounded in case of using 1 tx clone per 1 message.
